res fixed the font plexer which sometimes ignored scaling for the   first successfully loaded entry of a substitution sequences. res made iFontServer->Load() accept floats as font sizes. res tweaked a few FT2 warning messages. res fixed csfont to consider empty alpha data as "glyph not present".
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20849 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/font/server/csfont/csfont.h
@@ -113,7 +113,7 @@ public:
    * Query current font size. If server does not support rescalable
    * fonts, this method returns 0.
    */
-  virtual int GetSize ();
+  virtual float GetSize ();
 
   /**
    * Return the maximum width and height of a single glyph.
@@ -229,7 +229,7 @@ public:
    * Returns a new iFont object or 0 on failure.
    */
   virtual csPtr<iFont> LoadFont (const char *filename,
-    int size = 10);
+    float size = 10.0f);
 
   /// Called by child fonts to be added to font registry
   void NotifyCreate (csDefaultFont *font);
